refactor modificacion de carpetas para enums
