+++ src/main/java/dev/andreszapata/bankfuse/domain/enums/EstadoCuenta.java
@@ -1,0 +1,4 @@
+package dev.andreszapata.bankfuse.domain.enums;
+
+public enum EstadoCuenta {
+}
