+++ src/main/java/dev/andreszapata/bankfuse/domain/enums/TipoCuenta.java
@@ -1,0 +1,4 @@
+package dev.andreszapata.bankfuse.domain.enums;
+
+public enum TipoCuenta {
+}
