+++ src/main/java/dev/andreszapata/bankfuse/domain/enums/TipoTransaction.java
@@ -1,0 +1,4 @@
+package dev.andreszapata.bankfuse.domain.enums;
+
+public enum TipoTransaction {
+}
