+++ src/main/java/dev/andreszapata/bankfuse/infrastructure/db/springdata/entity/Transaction.java
@@ -1,6 +1,6 @@
 package dev.andreszapata.bankfuse.infrastructure.db.springdata.entity;
 
-import dev.andreszapata.bankfuse.infrastructure.db.springdata.enums.TipoTransaction;
+import dev.andreszapata.bankfuse.domain.enums.TipoTransaction;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
