--- src/main/java/dev/andreszapata/bankfuse/infrastructure/db/springdata/enums/TipoTransaction.java
@@ -1,4 +1,0 @@
-package dev.andreszapata.bankfuse.infrastructure.db.springdata.enums;
-
-public enum TipoTransaction {
-}
